Fix assumption that Huffman codes are 8-bit Added test to compress enwik8, the first 100MB of the English Wikipedia. This file is particularly interesting because it results in Huffman codes that are 20 bits in length. This is much larger than the 8 bits the HuffmanCoder could previously handle. The prefix table had to be made much larger - it must be 2^(max_code_length+1). The storage for Huffman codes must be larger.
Most of the changes are to the BitStream classes, to change the WriteBits and
ReadBits interfaces to support data types larger than 8-bits.

Worst-case, a Huffman code length could equal the number of symbols (256). Any
codes of that length will again break the HuffmanCoder. Also in general, large
code length hurts decoding performance. So a next-step will be to length-limit
the Huffman codes.

+++ fly/coders/huffman_types.h
@@ -9,7 +9,8 @@ namespace fly {
 typedef std::uint8_t symbol_type;
 typedef std::uint64_t frequency_type;
 
-typedef std::uint8_t code_type;
+// TODO: Once length limting is implemented, code_type can be 16-bit.
+typedef std::uint32_t code_type;
 typedef std::uint8_t length_type;
 
 struct HuffmanNode;
